requirements.txt
streamlit
openai
pandas
tabula-py
